pyproject.toml
[tool.poetry]
name = "expense-report"
version = "0.1.0"
description = ""
authors = ["gbieli <gery.bieli@gmx.ch>"]
readme = "README.md"
packages = [{include = "expense_report", from = "src"}]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.12"
tabula-py = "2.9.3"
openpyxl = "3.1.4"
pandas = "2.2.2"
JPype1 = "1.5.0"
pypdf = "^4.2.0"
ttp = "^0.9.5"
loguru = "^0.7.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"
bandit = "^1.7.9"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"

[tool.poetry.group.audit.dependencies]
licensecheck = "^2024.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
